Move client to pkg/
