+++ scanner/pkg/client/client.go
